Demostrating difference between Eyes Check Window and Eyes Check with Target Window
Added robot script to show the difference between the two methods.

An observation noting that the names are different since we run only one
visual check the different named checks are compared against the same
baseline. It appears it is checking based upon order and not name. ?If
we swapped two visual checks at some point will they be properly
checked?